fixed minor bugs in last change

+++ src/util/findglare.c
@@ -330,7 +330,7 @@ int	x, y;
 		xo = x-hl;
 	else
 		xo = 0;
-	disc = 1. - (xo*xo + y*y)/(sampdens*sampdens);
+	disc = 1. - (double)(xo*xo + y*y)/(sampdens*sampdens);
 	if (disc <= FTINY)
 		return(0.);
 	return(1./(sampdens*sampdens*sqrt(disc)));
